CB-27820 Only set /tmp to noexec if runtime version is 7.3.2 or later

+++ saltstack/base/salt/prerequisites/usr/bin/user-data-helper.sh
@@ -7,6 +7,8 @@ if [[ "$SECRET_ENCRYPTION_ENABLED" == "true" ]]; then
   source /usr/bin/cdp-retrieve-userdata-secrets.sh &> /var/log/cdp-retrieve-userdata-secrets.log
 fi
 
+function version { echo "$@" | awk -F. '{ printf("%d%03d%03d%03d\n", $1,$2,$3,$4); }'; }
+
 : ${CLOUD_PLATFORM:? required}
 : ${START_LABEL:? required}
 : ${PLATFORM_DISK_PREFIX:? required}
@@ -27,6 +29,7 @@ export IS_FREEIPA=true
 export IS_FREEIPA=false
 {% endif %}
 OS={{ pillar['OS'] }}
+STACK_VERSION={{ pillar['STACK_VERSION'] }}
 
 source /usr/bin/ccmv2-helper.sh
 
@@ -263,7 +266,7 @@ resize_partitions() {
       # Extend root logical volume to remaining free space
       lvextend -l +100%free -r /dev/mapper/rootvg-rootlv
     fi
-  else
+  elif [ -n "$STACK_VERSION" ] && [ $(version $STACK_VERSION) -gt $(version "7.3.1") ]; then
     # create and mount loopback filesystem for /tmp with same size as Azure logical volume
     dd if=/dev/zero of=/var/tmpfs bs=1M count=12288
     yes | mkfs.ext4 /var/tmpfs
